third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/only-shadow-host-in-shadow-tree.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../shadow/resources/shadow-dom.js"></script>
</head>
<body>
    <div id="sandbox"></div>
    <pre id="console"></pre>
</body>
<script>
description("crbug.com/359854: test for traversing elements in shadow tree.");

document.getElementById("sandbox").appendChild(
    createDOM("div", {"id": "host"},
        createShadowRoot(
            createDOM("div", {"id": "host2"},
                createShadowRoot(
                    createDOM("div", {})))),
        createDOM("div", {"id": "target"})));

shouldBe("document.querySelector('body /deep/ #target').id", '"target"');
</script>
</html>
